Week 5/requirements.txt
sqlalchemy
python-decouple
psycopg2
pandas
postgres
datetime
fastapi
uvicorn
Postman